+++ browser.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+	env: { browser: true },
+	parserOptions: { sourceType: 'module' },
+	rules: require('./rules/browser')
+};