+++ node.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+	env: { node: true },
+	rules: require('./rules/node')
+};