+++ rules/browser.js
@@ -1,0 +1,20 @@
+'use strict';
+
+module.exports = {
+	// Possible Errors
+	'no-console': 'warn',
+
+	// Best Practices
+	'no-alert': 'error',
+	'no-script-url': 'error',
+
+	// plugin:unicorn
+	'unicorn/prefer-add-event-listener': 'error',
+	'unicorn/prefer-dataset': 'error',
+	'unicorn/prefer-event-key': 'error',
+	'unicorn/prefer-modern-dom-apis': 'error',
+	'unicorn/prefer-node-append': 'error',
+	'unicorn/prefer-node-remove': 'error',
+	'unicorn/prefer-query-selector': 'error',
+	'unicorn/prefer-text-content': 'error',
+};