+++ rules/node.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+	// plugin:unicorn
+	'unicorn/no-new-buffer': 'error',
+	'unicorn/no-process-exit': 'error',
+};